+++ src/main/java/me/muksc/tacztweaks/mixin/feature/balancing/explosion_damage/EntityKineticBulletMixin.java
@@ -1,19 +1,11 @@
 package me.muksc.tacztweaks.mixin.feature.balancing.explosion_damage;
 
 import com.tacz.guns.entity.EntityKineticBullet;
-import com.tacz.guns.resource.pojo.data.gun.BulletData;
-import com.tacz.guns.resource.pojo.data.gun.GunData;
 import me.muksc.tacztweaks.config.Config;
-//~ if >=1.21.11 'ResourceLocation' -> 'Identifier'
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.Mth;
-import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.projectile.Projectile;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.Level;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
@@ -22,10 +14,18 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public abstract class EntityKineticBulletMixin {
     @Shadow private float explosionDamage;
 
-    //~ if >=1.21.11 'ResourceLocation' -> 'Identifier'
-    @Inject(method = "<init>(Lnet/minecraft/world/entity/EntityType;Lnet/minecraft/world/level/Level;Lnet/minecraft/world/entity/LivingEntity;Lnet/minecraft/world/item/ItemStack;Lnet/minecraft/resources/ResourceLocation;Lnet/minecraft/resources/ResourceLocation;Lnet/minecraft/resources/ResourceLocation;ZLcom/tacz/guns/resource/pojo/data/gun/GunData;Lcom/tacz/guns/resource/pojo/data/gun/BulletData;)V", at = @At("TAIL"))
-    //~ if >=1.21.11 'ResourceLocation' -> 'Identifier'
-    private void tacztweaks$init$explosionDamageModifier(EntityType<? extends Projectile> type, Level level, LivingEntity shooter, ItemStack gunItem, ResourceLocation ammoId, ResourceLocation gunId, ResourceLocation displayId, boolean tracer, GunData gunData, BulletData bulletData, CallbackInfo ci) {
-        explosionDamage = Mth.clamp((float) Config.Balancing.ExplosionDamage.eval(explosionDamage), 0.0F, Float.MAX_VALUE);
+    @Unique
+    private boolean tacztweaks$explosionDamageApplied;
+
+    @Inject(method = "<init>", at = @At("RETURN"), remap = false)
+    private void tacztweaks$init$explosionDamageModifier(CallbackInfo ci) {
+        if (tacztweaks$explosionDamageApplied) return;
+        if (((EntityKineticBullet) (Object) this).getOwner() == null) return;
+        tacztweaks$explosionDamageApplied = true;
+        explosionDamage = Mth.clamp(
+            (float) Config.Balancing.ExplosionDamage.eval(explosionDamage),
+            0.0F,
+            Float.MAX_VALUE
+        );
     }
 }