+++ src/main/java/me/muksc/tacztweaks/mixin/feature/general/fixes/disable_bullet_culling/EntityBulletRendererMixin.java
@@ -11,7 +11,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(value = EntityBulletRenderer.class, remap = false)
 public abstract class EntityBulletRendererMixin {
-    @Inject(method = "shouldRender(Lcom/tacz/guns/entity/EntityKineticBullet;Lnet/minecraft/client/renderer/culling/Frustum;DDD)Z", at = @At("HEAD"), cancellable = true)
+    @Inject(method = "shouldRender", at = @At("HEAD"), cancellable = true)
     private void tacztweaks$shouldRender$disableBulletCulling(EntityKineticBullet bullet, Frustum camera, double pCamX, double pCamY, double pCamZ, CallbackInfoReturnable<Boolean> cir) {
         if (Config.General.Fixes.disableBulletCulling()) cir.setReturnValue(true);
     }